small fixes

+++ application/statistics/views.py
@@ -12,7 +12,7 @@ def statistics_show():
 	correctAnswers = UsersChoice.countCorrectAnswers(current_user.id)
 	percentageOfCorrectAnswers = 0
 	if allAnswers > 0:
-		percentageOfCorrectAnswers = correctAnswers/allAnswers*100
+		percentageOfCorrectAnswers = int(correctAnswers/allAnswers*100)
 
 	categories = Question.findAllCategoriesInUse()
 
@@ -23,7 +23,7 @@ def statistics_show():
 		categoryCorrect = UsersChoice.countCorrectAnswersByCategory(current_user.id, c)
 		categoryPercentage = 0
 		if categoryAnswers > 0:
-			categoryPercentage = categoryCorrect/categoryAnswers*100
+			categoryPercentage = int(categoryCorrect/categoryAnswers*100)
 		answersByCategories.append({"category":category, "answers":categoryAnswers, "correct":categoryCorrect, "percentage":categoryPercentage})
 
 	return render_template("statistics/statistics.html", allAnswers = allAnswers, correctAnswers = correctAnswers, percentageOfCorrectAnswers = percentageOfCorrectAnswers, answersByCategories = answersByCategories)
